skills/pm-skills/pm-workflow/references/pm-knowledge/scripts/requirements.txt
pyyaml>=6.0
pytest>=8.0